KA-281 Geting data from server is working
Create new getItemsFactory that dispatches success or failure based on getItems server result. The fetch service, success function and failure function are provided by the DI. Server data are changed to client data and reducer method is called.
Also add test to factory and to the reducer method.
If failed, no functionality yet.